+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Latest
 
+- adds conditional formatting
 # Release 0.2.0
 
 Features: